requirements.txt
redis
pyyaml~=6.0
selenium~=4.1.0
jsonpath
pymysql
webdriver-helper
rsa~=4.8
openpyxl~=3.0.9
flask
allure-pytest
requests
pytest~=6.2.5
pytest-base-url
pytest-rerunfailures
pytest-ordering
pytest-html
pytest-xdist
ddt~=1.4.4